Al sintonizar un canal HD se mostraba el último frame del canal anterior, por lo que de momento es mejor que los detecte como canales de radio.
Explicación del código fuente:
This more than ugly but driver is more than ugly as bad.
If switched to a radio channel, driver does not blank screen.
Last frame is shown until another video-channel is selected.
Even worse if started up with a radio channel video gets not
shown if a video channel is selected.
Have I mentioned the driver is buggy as hell, yet?
This is the only way I found to get it blanking